requirements.txt
aiogram==3.11.0
python-dotenv==1.0.0